+++ Larvel/database/migrations/2021_07_01_154731_create__p_d_c_pphoto.php
@@ -16,7 +16,7 @@ class CreatePDCPphoto extends Migration
         Schema::create('photos', function (Blueprint $table) {
             $table->increments('id');
             $table->text('path');
-            $table->unsignedInteger('program_id');	
+            $table->unsignedBigInteger('program_id');
             $table->foreign('program_id')->references('id')->on('programs')->onDelete('cascade');
             $table->timestamps();
         });
