+++ Larvel/database/migrations/2021_07_01_154837_create__p_d_c_pagenda.php
@@ -16,7 +16,7 @@ class CreatePDCPagenda extends Migration
         Schema::create('agendas', function (Blueprint $table) {
             $table->increments('id');
             $table->longText('agenda');
-            $table->unsignedInteger('program_id');	
+            $table->unsignedBigInteger('program_id');
             $table->foreign('program_id')->references('id')->on('programs')->onDelete('cascade');
             $table->timestamps();
         });
