+++ Larvel/database/migrations/2021_07_01_160919_create__p_d_c_feedback.php
@@ -16,7 +16,7 @@ class CreatePDCFeedback extends Migration
         Schema::create('feedbacks', function (Blueprint $table) {
             $table->increments('id');
             // $table->longText('comment')->nullable();
-            $table->unsignedInteger('program_id');
+            $table->unsignedBigInteger('program_id');
             $table->foreign('program_id')->references('id')->on('programs')->onDelete('cascade');
             $table->timestamps();
         });
